+++ src/cache/QueryResultCacheFactory.ts
@@ -27,7 +27,10 @@ export class QueryResultCacheFactory {
             throw new Error(`To use cache you need to enable it in connection options by setting cache: true or providing some caching options. Example: { host: ..., username: ..., cache: true }`);
 
         if ((this.connection.options.cache as any).type === "redis")
-            return new RedisQueryResultCache(this.connection);
+            return new RedisQueryResultCache(this.connection, "redis");
+
+        if ((this.connection.options.cache as any).type === "ioredis")
+            return new RedisQueryResultCache(this.connection, "ioredis");
 
         return new DbQueryResultCache(this.connection);
     }