+++ src/connection/options-reader/ConnectionOptionsEnvReader.ts
@@ -33,6 +33,7 @@ export class ConnectionOptionsEnvReader {
             migrationsRun: OrmUtils.toBoolean(PlatformTools.getEnvVariable("TYPEORM_MIGRATIONS_RUN")),
             entities: this.stringToArray(PlatformTools.getEnvVariable("TYPEORM_ENTITIES")),
             migrations: this.stringToArray(PlatformTools.getEnvVariable("TYPEORM_MIGRATIONS")),
+            migrationsTableName: PlatformTools.getEnvVariable("TYPEORM_MIGRATIONS_TABLE_NAME"),
             subscribers: this.stringToArray(PlatformTools.getEnvVariable("TYPEORM_SUBSCRIBERS")),
             logging: this.transformLogging(PlatformTools.getEnvVariable("TYPEORM_LOGGING")),
             logger: PlatformTools.getEnvVariable("TYPEORM_LOGGER"),
@@ -43,7 +44,8 @@ export class ConnectionOptionsEnvReader {
                 entitiesDir: PlatformTools.getEnvVariable("TYPEORM_ENTITIES_DIR"),
                 migrationsDir: PlatformTools.getEnvVariable("TYPEORM_MIGRATIONS_DIR"),
                 subscribersDir: PlatformTools.getEnvVariable("TYPEORM_SUBSCRIBERS_DIR"),
-            }
+            },
+            cache: this.transformCaching()
         };
     }
 
@@ -61,6 +63,26 @@ export class ConnectionOptionsEnvReader {
             return "all";
 
         return this.stringToArray(logging);
+    }
+
+    /**
+     * Transforms caching option into real caching value option requires.
+     */
+    protected transformCaching(): boolean | object | undefined {
+        const caching = PlatformTools.getEnvVariable("TYPEORM_CACHE");
+        if (caching === "true" || caching === "TRUE" || caching === "1")
+            return true;
+        if (caching === "false" || caching === "FALSE" || caching === "0")
+            return false;
+        if (caching === "redis" || caching === "database")
+            return {
+                type: caching,
+                options: PlatformTools.getEnvVariable("TYPEORM_CACHE_OPTIONS") ? JSON.parse(PlatformTools.getEnvVariable("TYPEORM_CACHE_OPTIONS")) : undefined,
+                alwaysEnabled: PlatformTools.getEnvVariable("TYPEORM_CACHE_ALWAYS_ENABLED"),
+                duration: parseInt(PlatformTools.getEnvVariable("TYPEORM_CACHE_DURATION"))
+            };
+
+        return undefined;
     }
 
     /**