+++ src/driver/postgres/PostgresDriver.ts
@@ -599,6 +599,9 @@ export class PostgresDriver implements Driver {
         } else if (typeof defaultValue === "string") {
             return `'${defaultValue}'${arrayCast}`;
 
+        } else if (defaultValue === null) {
+            return `null`;
+
         } else if (typeof defaultValue === "object") {
             return `'${JSON.stringify(defaultValue)}'`;
 
@@ -726,7 +729,7 @@ export class PostgresDriver implements Driver {
                 || tableColumn.precision !== columnMetadata.precision
                 || tableColumn.scale !== columnMetadata.scale
                 // || tableColumn.comment !== columnMetadata.comment // todo
-                || (!tableColumn.isGenerated && this.normalizeDefault(columnMetadata) !== tableColumn.default) // we included check for generated here, because generated columns already can have default values
+                || (!tableColumn.isGenerated && this.lowerDefaultValueIfNessesary(this.normalizeDefault(columnMetadata)) !== tableColumn.default) // we included check for generated here, because generated columns already can have default values
                 || tableColumn.isPrimary !== columnMetadata.isPrimary
                 || tableColumn.isNullable !== columnMetadata.isNullable
                 || tableColumn.isUnique !== this.normalizeIsUnique(columnMetadata)
@@ -736,7 +739,15 @@ export class PostgresDriver implements Driver {
                 || tableColumn.srid !== columnMetadata.srid;
         });
     }
-
+    private lowerDefaultValueIfNessesary(value: string | undefined) {
+        // Postgres saves function calls in default value as lowercase #2733
+        if (!value) {
+            return value;
+        }
+        return value.split(`'`).map((v, i) => {
+            return i % 2 === 1 ? v : v.toLowerCase();
+        }).join(`'`);
+    }
     /**
      * Returns true if driver supports RETURNING / OUTPUT statement.
      */